fix(server): use bool directly for ReadDigitalIO response value
The proto field ReadDigitalIoResponse.value is bool, not u32.
Remove the u32::from() conversion.

+++ crates/server/src/grpc/ni_daq_service.rs
@@ -953,7 +953,7 @@ impl NiDaqService for NiDaqServiceImpl {
                     s.get_value(&pin.to_string()).await
                 })
                 .await?;
-            let value = u32::from(json_value.as_bool().unwrap_or(false));
+            let value = json_value.as_bool().unwrap_or(false);
 
             Ok(Response::new(ReadDigitalIoResponse {
                 success: true,
